Changing the demo from mvg to code.
THIS CHANGE IS NOT FOR GENERAL RELEASE OR PUBLIC CONSUMPTION.
It is just a scratch file for bits of code that might later get
incorporated into the project.

The mvg file for the chronodex can now be generated as
$ ruby bin/gen_chronodex > test.mvg
$ convert -background none test.mvg test.png
$ open test.png

This bit of code should be useful for generating the overlay; in
particular the inner radius (6am to 9am) now has code to generate
it, as does the 9pm to 12am dashed arc.